# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/technologies": {
            "get": {
                "tags": [
                    "Technologies"
                ],
                "summary": "Get all technologies",
                "operationId": "b40b9b0c7f554136d29280603dc43d37",
                "responses": {
                    "200": {
                        "description": "List of technologies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "icon": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Technologies"
                ],
                "summary": "Create a new technology",
                "operationId": "f5a8c65fdc834222d05ba9d4178e0ee9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "icon": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Technology created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "icon": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/technologies/{id}": {
            "get": {
                "tags": [
                    "Technologies"
                ],
                "summary": "Get a technology by ID",
                "operationId": "f270d6824b4d684e9d33905033e0b2f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the technology",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Technology retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "icon": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Technology not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Technologies"
                ],
                "summary": "Update a technology",
                "operationId": "39f194393e7a23886ce98366f92d8bda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the technology to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "icon": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Technology updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "icon": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Technology not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Technologies"
                ],
                "summary": "Delete a technology",
                "operationId": "2584c343ccb043e728c2a514ac373f82",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the technology to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Technology deleted successfully"
                    },
                    "404": {
                        "description": "Technology not found"
                    }
                }
            }
        },
        "/api/translations/{id}": {
            "delete": {
                "tags": [
                    "Translations"
                ],
                "summary": "Delete a translation",
                "operationId": "d0d0d6669d70520f4119d706d6567816",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the translation to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Translation deleted successfully"
                    },
                    "404": {
                        "description": "Translation not found"
                    }
                }
            }
        }
    },
...
}